+++ src/types/zeta.ts
@@ -5031,6 +5031,23 @@ export type Zeta = {
       }
     },
     {
+      "name": "OrderCompleteType",
+      "type": {
+        "kind": "enum",
+        "variants": [
+          {
+            "name": "Cancel"
+          },
+          {
+            "name": "Fill"
+          },
+          {
+            "name": "Booted"
+          }
+        ]
+      }
+    },
+    {
       "name": "MarginRequirement",
       "type": {
         "kind": "enum",
@@ -5197,7 +5214,7 @@ export type Zeta = {
       ]
     },
     {
-      "name": "CancelEvent",
+      "name": "OrderCompleteEvent",
       "fields": [
         {
           "name": "marginAccount",
@@ -5229,7 +5246,7 @@ export type Zeta = {
           "index": false
         },
         {
-          "name": "size",
+          "name": "unfilledSize",
           "type": "u64",
           "index": false
         },
@@ -5241,6 +5258,13 @@ export type Zeta = {
         {
           "name": "clientOrderId",
           "type": "u64",
+          "index": false
+        },
+        {
+          "name": "orderCompleteType",
+          "type": {
+            "defined": "OrderCompleteType"
+          },
           "index": false
         }
       ]
@@ -10838,6 +10862,23 @@ export const IDL: Zeta = {
       }
     },
     {
+      "name": "OrderCompleteType",
+      "type": {
+        "kind": "enum",
+        "variants": [
+          {
+            "name": "Cancel"
+          },
+          {
+            "name": "Fill"
+          },
+          {
+            "name": "Booted"
+          }
+        ]
+      }
+    },
+    {
       "name": "MarginRequirement",
       "type": {
         "kind": "enum",
@@ -11004,7 +11045,7 @@ export const IDL: Zeta = {
       ]
     },
     {
-      "name": "CancelEvent",
+      "name": "OrderCompleteEvent",
       "fields": [
         {
           "name": "marginAccount",
@@ -11036,7 +11077,7 @@ export const IDL: Zeta = {
           "index": false
         },
         {
-          "name": "size",
+          "name": "unfilledSize",
           "type": "u64",
           "index": false
         },
@@ -11048,6 +11089,13 @@ export const IDL: Zeta = {
         {
           "name": "clientOrderId",
           "type": "u64",
+          "index": false
+        },
+        {
+          "name": "orderCompleteType",
+          "type": {
+            "defined": "OrderCompleteType"
+          },
           "index": false
         }
       ]